oceanics-io-api/requirements.txt
-e .
attrs==21.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
decorator==4.4.2
flask-cors==3.0.10
flask==1.1.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==1.7.0; python_version < '3.8'
inflection==0.5.1
itsdangerous==1.1.0
jinja2==2.11.3
jsonschema==3.2.0
mako==1.1.4
markdown==3.3.3
markupsafe==1.1.1
minio==7.0.2
neo4j==4.2.1
openapi-spec-validator==0.2.9
passlib==1.7.2
pdoc3==0.9.2
prance==0.19.0
py==1.10.0
pyrsistent==0.17.3
pytz==2020.1
pyyaml==5.4.1
requests==2.24.0
retry==0.9.2
semver==2.13.0
six==1.15.0
urllib3==1.25.11
werkzeug==1.0.1
wheel==0.37.0
zipp==3.4.0
